A Quick Commit

+++ core/src/test/java/com/marmoush/jutils/core/adapter/generator/id/IdGeneratorsTest.java
@@ -5,15 +5,16 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class IdGeneratorsTest {
   @Test
   public void SerialIdTest() {
-    IdGenerator idGen = new SerialIdGenerator(new AtomicInteger());
+    IdGenerator idGen = new SerialIdGenerator(new AtomicLong());
     Assertions.assertEquals("0", idGen.generate());
     Assertions.assertEquals("1", idGen.generate());
     Assertions.assertEquals("2", idGen.generate());
-    idGen = new SerialIdGenerator(new AtomicInteger());
+    idGen = new SerialIdGenerator(new AtomicLong());
     Assertions.assertEquals("0", idGen.generate());
   }
